A lot of large and small changes that make the new GeoJSON vector layer fully dynamic
As I did testing on the contributed layer, I ended up finding a lot of misc. issues with making the layer fully dynamic. All of the px-map layers try to respect the idea that failure is "OK" -- the developer might pass in a bad attribute, might remove an attribute all together and detach the layer only to re-use it later, or might completely change a layer's behavior at any time. This requires all APIs to respond well regardless of what is given to them (e.g. data going from defined to null to defined). These

The changes I made to the `getInstOptions` and `updateInst` flows (and accompanying tests) ensure that any attribute can recieve almost any valid value and react well, avoiding fatal TypeErrors or other hard runtime errors the layer cannot recover from.

The contribution was originally named `px-map-vector-layer`. To maintain consistency with naming, it seemed like a good idea to instead call it a GeoJSON layer (since that's the kind of data it accepts/draws) and make it a sub-category of layer components. I renamed it `px-map-layer-geojson`. That seems to match well with `px-map-marker-static`, `px-map-marker-symbol`, etc. This will now be grouped with `px-map-layer-group` and other soon-to-come layer types. (Aside: the only unfortunate outlier here is `px-map-tile-layer`, which is not a sibling/kin of `px-map-layer-*` components. It's a base layer, not an overlay -- it can only be used alone. The layer kin can be added on top of one another as overlays. Perhaps that was a bad name for the tile layer component tag. Oh well.)

We're now sending events out of the web component for lifecycle processes (e.g. when a feature is added as a layer/removed as a layer) and interaction (e.g. when the user clicks on a feature, or opens a feature popup).

I removed the `blankLayer` API. Instead, we follow the pattern established in other px-map components: if `canAddInst()` returns false because needed attributes (in this case `data`) are not defined, we wait until a change is made to `data` and run the `canAddInst()` -> `createInst()` chain of logic again. We keep checking and running until we `canAddInst()`.

This also adds more tests and more examples to cover all of the above.

This contribution was carried forward by their great team. Thanks for helping us build more useful px-map features.