Patches popups to ensure they are measured correctly in shady DOM
Oh, sad days. If the map is loaded in shady DOM, and the user clicks on a marker to open a popup, Leaflet will try to measure the popup's `clientHeight` to determine if it needs to pan the map to keep the popup in view. Unfortunately, Leaflet does these measurements before Polymer has a chance to scope the popup's subtree. For data popups, this was creating a condition where Leaflet thought the popup was far bigger than it actually was, and panned the map too far, causing janky interaction.

To fix this, I held my nose a bit and monkey patched some HTMLElements as their `appendChild` and `innerHTML` methods/setters were used inside Leaflet. This is not a great solution. It's not good. But I don't see many other options without ripping Leaflet apart. This monkey patch only happens when shady DOM is enabled -- we skip it and just let sane things happen in shadow DOM.