Wait to create Leaflet instance until after first render in Polymer 2.x
In Polymer 1.x the px-map.html element's template was created relatively
early in the element lifecycle. By the time the code to create the
L.map instance was run, the px-map's template had stamped its template
and had time to render the container element that would hold the map.
Timing changed in Polymer 2.x and the attached lifecycle is run earlier.

When Leaflet's L.map constructor is called and passed a container
element that hasn't painted yet, it can lead to weird outcomes like
GeoJSON and other projected points not drawing in the correct location
on the map. In Polymer 2.x world, we force ourselves to wait until
afterNextRender in attached to ensure our map container has had
time to paint and prevent this issue.