Remove for...of iterators that are transpiled to use Symbols
The polymer build toolchain transpiles for...of loops down to ES5
using traditional for loops and Symbols. IE11 does not have native
support for Symbols. If the component or app does not load a
runtime Symbol polyfill, IE11 will throw an error when it runs
this transpiled code. Switched out to use regular old forEach
and for loops to solve this problem.